fix other issues with ZodType
