+++ src/__tests__/refine.test.ts
@@ -56,8 +56,8 @@ test("refinement type guard", () => {
   type Input = z.input<typeof validationSchema>;
   type Schema = z.infer<typeof validationSchema>;
 
-  util.assertEqual<"a", Input["a"]>(false);
-  util.assertEqual<string, Input["a"]>(true);
+  util.assertEqual<"a", Input["a"]>(true);
+  util.assertEqual<string, Input["a"]>(false);
 
   util.assertEqual<"a", Schema["a"]>(true);
   util.assertEqual<string, Schema["a"]>(false);
@@ -274,25 +274,4 @@ test("fatal superRefine", () => {
 
   expect(result.success).toEqual(false);
   if (!result.success) expect(result.error.issues.length).toEqual(1);
-});
-
-test("compatibility with ZodType", () => {
-  const stringIsValid = (value: string): value is "a" | "b" =>
-    ["a", "b"].includes(value);
-
-  const fooSchema = z.object({
-    type: z.string().refine(stringIsValid),
-  });
-
-  type Foo = z.infer<typeof fooSchema>;
-
-  type Bar = {
-    children: Foo[];
-  };
-
-  const barSchema = z.object({
-    children: fooSchema.array(),
-  });
-
-  util.assertIs<z.ZodType<Bar>>(barSchema);
 });
